BackEnd/requirements.txt
Django==5.2.10
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt